requirements.txt
pytesseract
Pillow
opencv-python
fastapi
pydantic
python-dotenv
openai
requests
pandas
numpy
beautifulsoup4
playwright
uvicorn